# openapi.yaml
openapi: 3.0.1
info:
  title: Smart Crop Advisory System - API
  version: 1.0.0
  description: API for smart-crop-advisory-system
servers:
  - url: http://localhost:4000/api/v1
paths:
  /auth/signup:
    post:
      summary: Signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                password: { type: string }
      responses:
        '201': { description: Created }
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200': 
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken: { type: string }
  /farms:
    post:
      summary: Create farm
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                location: { type: object }
  /soil/{farmId}:
    post:
      summary: Push soil reading
      parameters:
        - name: farmId
          in: path
          required: true
          schema: { type: string }
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pH: { type: number }
                moisture: { type: number }
      responses:
        '201': { description: Created }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT